# docs/openapi.yml
openapi: "3.0.3"
info:
  title: 云裳衣裳 — 时尚推荐平台 API
  description: |
    AI 驱动的智能时尚穿搭推荐平台。支持图片上传、CLIP 风格分析、
    天气感知推荐、知识库 RAG 对话和衣橱管理。
  version: "1.0.0"
  contact:
    name: Fashion Platform Team
    url: https://github.com/your-org/fashion-platform
servers:
  - url: http://localhost:5000
    description: 本地开发服务器
...
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "错误描述"
      required: [success, error]
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie
  responses:
...
    Unauthorized:
      description: 需要登录
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 用户注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
              required: [username, email, password]
      responses:
        "200":
          description: 注册成功
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: 登录成功
        "401":
          description: 认证失败
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      operationId: getCurrentUser
      security:
        - sessionAuth: []
      responses:
        "200":
          description: 当前用户 JSON
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录 (API)
      operationId: apiLogout
      security:
        - sessionAuth: []
      responses:
        "200":
          description: 退出成功

  # ==================== Recommendation ====================
...